requirements.txt
django
requests
psycopg2-binary
python-dotenv